LayoutTests/media/media-source/media-source-real-rapid-appends.html

<!DOCTYPE html>
<html>
<head>
    <title>media-source-real-rapid-appends</title>
    <script src="mp4-generator.js"></script>
    <script src="../video-test.js"></script>
    <script>
    // Tests appending many short segments sequentially, simulating adaptive
    // streaming behavior. This exercises the IPC pipeline under load — a
    // pattern that exposed back-pressure and CPU-pegging bugs during the
    // re-architecture (where each sample dispatched as a separate main-thread task).

    var source;
    var sourceBuffer;

    async function runTest() {
        findMediaElement();

        source = new MediaSource();
        video.src = URL.createObjectURL(source);
        await waitFor(source, 'sourceopen');

        sourceBuffer = source.addSourceBuffer(MP4.VIDEO_TYPE);

        var initData = MP4.initSegment({
            timescale: 1000,
            tracks: [{ id: 1, type: 'video', timescale: 1000 }]
        });
        sourceBuffer.appendBuffer(initData);
        await waitFor(sourceBuffer, 'updateend');

        // Append 10 one-second segments sequentially
        for (var i = 0; i < 10; i++) {
            var media = MP4.mediaSegment({
                sequenceNumber: i + 1,
                tracks: [{ id: 1, type: 'video', baseDecodeTime: i * 1000, samples: [
                    { duration: 1000, compositionTimeOffset: 0, isSync: true },
                ]}]
            });
            sourceBuffer.appendBuffer(media);
            await waitFor(sourceBuffer, 'updateend');
        }

        // Should have one contiguous range [0, 10)
        testExpected('sourceBuffer.buffered.length', 1);
        testExpectedEqualWithTolerance('sourceBuffer.buffered.start(0)', 0, 0.01);
        testExpectedEqualWithTolerance('sourceBuffer.buffered.end(0)', 10, 0.01);

        run('source.endOfStream()');
        testExpected('source.readyState', 'ended');
        testExpectedEqualWithTolerance('source.duration', 10, 0.01);

        endTest();
    }
    </script>
</head>
<body onload="runTest()">
    <video></video>
</body>
</html>
